Move fixtures into conftest.py
