+++ tests/benzene.mol
