--- tests/fixtures/__init__.py
@@ -1,1 +1,0 @@
-from .benzene import *  # noqa
